changed amount types from float to double

+++ src/main/java/fr/flaurens/bankaccount/model/Account.java
@@ -4,18 +4,18 @@ public class Account {
 
     private final long accountId;
 
-    private float balance;
+    private double balance;
 
     public Account(long accountId){
         this.accountId = accountId;
         this.balance = 0f;
     }
 
-    public float getBalance(){
+    public double getBalance(){
         return this.balance;
     }
 
-    public float updateBalance(float amount){
+    public double updateBalance(double amount){
         this.balance += amount;
         return this.balance;
     }
